examples/adk-a2a-agents/requirements.txt
google-adk>=1.0.0
a2ui-agent>=0.9.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
